+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Compressed results of `getAllTagGroups` to reduce tool output size.
 - Compressed results of `getAllEntries` to reduce tool output size.
 - Compressed results of `getAllSections` to reduce tool output size.
+- Compressed results of `getAllEntryTypes` to reduce tool output size.
 
 ### Fixed
 - Fixed console error.